Dictionary logic updates
Changed the dictionaries to use a for loop instead of popping. This change also happens in the Merge() function. I was going to change the variable names near line 344, but I did not know if it would be appropriate, just to make it like the previous for loop. I also copied the function from the previous for loop to determine if running a chi-square test is appropriate. Also, in line 118, there are multiple definitions of the extreme directory and I am not sure if you would like to remove one.